+++ pom.xml
@@ -148,8 +148,7 @@
 				<artifactId>maven-compiler-plugin</artifactId>
 				<version>3.8.1</version>
 				<configuration>
-					<source>11</source>
-					<target>11</target>
+					<release>11</release>
 					<showDeprecation>true</showDeprecation>
 					<showWarnings>true</showWarnings>
 					<useIncrementalCompilation>false</useIncrementalCompilation>
@@ -158,8 +157,6 @@
 						<arg>-Xlint:all</arg>
 						<arg>-Xlint:-requires-automatic</arg>
 						<arg>-Xpkginfo:always</arg>
-						<arg>--add-exports</arg>
-						<arg>java.base/jdk.internal.misc=hosh</arg>
 					</compilerArgs>
 				</configuration>
 			</plugin>
@@ -170,9 +167,6 @@
 				<configuration>
 					<redirectTestOutputToFile>true</redirectTestOutputToFile>
 					<trimStackTrace>false</trimStackTrace>
-					<argLine>
-						--add-opens hosh.spi/hosh.spi=ALL-UNNAMED
-					</argLine>
 				</configuration>
 			</plugin>
 			<plugin>
