+++ runtime/pom.xml
@@ -3,14 +3,12 @@
 		 xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
 		 xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<parent>
-		<artifactId>hosh</artifactId>
 		<groupId>dfa1</groupId>
+		<artifactId>hosh</artifactId>
 		<version>0.1.0-SNAPSHOT</version>
 	</parent>
 	<modelVersion>4.0.0</modelVersion>
-
 	<artifactId>hosh-runtime</artifactId>
-
 	<dependencies>
 		<dependency>
 			<groupId>dfa1</groupId>
