--- runtime/src/main/java/hosh/BootstrapBuiltins.java
@@ -1,45 +1,0 @@
-/*
- * MIT License
- *
- * Copyright (c) 2018-2020 Davide Angelocola
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-package hosh;
-
-import hosh.runtime.SimpleCommandRegistry;
-import hosh.spi.CommandRegistry;
-import hosh.spi.Module;
-import hosh.spi.State;
-
-import java.util.ServiceLoader;
-
-/**
- * Registers all built-in command, used in both production and test.
- */
-public class BootstrapBuiltins {
-
-	public void registerAllBuiltins(State state) {
-		CommandRegistry commandRegistry = new SimpleCommandRegistry(state);
-		ServiceLoader<Module> modules = ServiceLoader.load(Module.class);
-		for (Module module : modules) {
-			module.initialize(commandRegistry);
-		}
-	}
-}
