+++ runtime/src/main/java/hosh/runtime/Injector.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package runtime;
+package hosh.runtime;
 
 import hosh.spi.Command;
 import hosh.spi.HistoryAware;
