+++ runtime/src/main/java/hosh/runtime/Interpreter.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package runtime;
+package hosh.runtime;
 
 import jfr.CommandExecution;
 import hosh.spi.Command;
