+++ runtime/src/main/java/hosh/runtime/InterpreterAware.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package runtime;
+package hosh.runtime;
 
 /**
  * Marker interface for special commands like @see {@link PipelineCommand} or
