+++ runtime/src/main/java/hosh/runtime/LambdaCommand.java
@@ -21,9 +21,10 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package runtime;
+package hosh.runtime;
 
-import hosh.doc.Todo;
+import hosh.spi.Todo;
+import hosh.runtime.Interpreter;
 import hosh.spi.Command;
 import hosh.spi.ExitStatus;
 import hosh.spi.InputChannel;
