+++ runtime/src/main/java/hosh/runtime/PathInitializer.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package hosh;
+package hosh.runtime;
 
 import java.io.File;
 import java.nio.file.Path;
