+++ runtime/src/main/java/hosh/runtime/VersionLoader.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package runtime;
+package hosh.runtime;
 
 import java.io.IOException;
 import java.io.InputStream;
