+++ runtime/src/test/java/hosh/PathInitializerTest.java
@@ -23,6 +23,7 @@
  */
 package hosh;
 
+import hosh.runtime.PathInitializer;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.condition.EnabledOnOs;
