+++ runtime/src/test/java/hosh/fitness/DocFitnessTest.java
@@ -30,8 +30,8 @@ import com.tngtech.archunit.lang.ArchCondition;
 import com.tngtech.archunit.lang.ArchRule;
 import com.tngtech.archunit.lang.ConditionEvents;
 import com.tngtech.archunit.lang.SimpleConditionEvent;
-import hosh.BootstrapBuiltins;
-import hosh.Hosh;
+import hosh.runtime.BootstrapBuiltins;
+import hosh.runtime.Hosh;
 import hosh.doc.Description;
 import hosh.doc.Example;
 import hosh.doc.Examples;
