+++ spi/pom.xml
@@ -8,7 +8,5 @@
 		<version>0.1.0-SNAPSHOT</version>
 	</parent>
 	<modelVersion>4.0.0</modelVersion>
-
 	<artifactId>hosh-spi</artifactId>
-
 </project>
